Support passing in file size to read stream
Saves 1 syscall for the zero-byte read() return.  Not worth statting
over, since that'd just add the syscall back in, but if it happens that
you have already run a stat on this file, and are confident it won't
change while reading, then it can be efficient to pass in the expected
size.